deps/v8/test/webkit/strict-callback-this-expected.txt

# Copyright 2013 the V8 project authors. All rights reserved.
# Copyright (C) 2005, 2006, 2007, 2008, 2009 Apple Inc. All rights reserved.
#
# Redistribution and use in source and binary forms, with or without
# modification, are permitted provided that the following conditions
# are met:
# 1.  Redistributions of source code must retain the above copyright
#     notice, this list of conditions and the following disclaimer.
# 2.  Redistributions in binary form must reproduce the above copyright
#     notice, this list of conditions and the following disclaimer in the
#     documentation and/or other materials provided with the distribution.
#
# THIS SOFTWARE IS PROVIDED BY APPLE INC. AND ITS CONTRIBUTORS ``AS IS'' AND ANY
# EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
# WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
# DISCLAIMED. IN NO EVENT SHALL APPLE INC. OR ITS CONTRIBUTORS BE LIABLE FOR ANY
# DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
# (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
# LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
# ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
# (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
# SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

This tests that a call to array/string prototype methods pass the correct this value (undefined) to strict callees.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testArrayPrototypeSort(strictThrowThisString) is undefinedString
PASS testArrayPrototypeFilter(strictThrowThisString) is undefinedString
PASS testArrayPrototypeMap(strictThrowThisString) is undefinedString
PASS testArrayPrototypeEvery(strictThrowThisString) is undefinedString
PASS testArrayPrototypeForEach(strictThrowThisString) is undefinedString
PASS testArrayPrototypeSome(strictThrowThisString) is undefinedString
PASS testStringPrototypeReplace(strictThrowThisString) is undefinedString
PASS testArrayPrototypeSort(nonstrictThrowThisString) is globalObjectString
PASS testArrayPrototypeFilter(nonstrictThrowThisString) is globalObjectString
PASS testArrayPrototypeMap(nonstrictThrowThisString) is globalObjectString
PASS testArrayPrototypeEvery(nonstrictThrowThisString) is globalObjectString
PASS testArrayPrototypeForEach(nonstrictThrowThisString) is globalObjectString
PASS testArrayPrototypeSome(nonstrictThrowThisString) is globalObjectString
PASS testStringPrototypeReplace(nonstrictThrowThisString) is globalObjectString
PASS successfullyParsed is true

TEST COMPLETE
